Rename and relocate ImportRoute
ImportRoute is not a route anymore.
